sm msg
