+++ package.json
@@ -18,7 +18,7 @@
   "devDependencies": {
     "@mate-academy/eslint-config-react": "*",
     "@mate-academy/eslint-config-react-typescript": "*",
-    "@mate-academy/scripts": "^0.9.0",
+    "@mate-academy/scripts": "^0.9.14",
     "@mate-academy/students-ts-config": "*",
     "@mate-academy/stylelint-config": "*",
     "cypress": "^5.6.0",
