+++ src/App.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { Person } from './components/Person';
+
+const misha = {
+  name: 'Misha',
+  age: 37,
+  sex: 'm',
+  isMarried: true,
+  partnerName: 'Natasha',
+};
+
+const olya = {
+  name: 'Olya',
+  sex: 'f',
+  isMarried: true,
+  partnerName: 'Maksym',
+};
+
+const alex = {
+  name: 'Alex',
+  age: 25,
+  sex: 'm',
+  isMarried: false,
+};
+
+const App = () => (
+  <div className="App">
+    <Person {...misha} />
+    <Person {...olya} />
+    <Person {...alex} />
+  </div>
+);
+
+export default App;
