+++ src/components/Person.scss
